Link updates, and table added
took out the footer, added in a table to the bottom of the body on home page. added a few html content to the other page links. edited the email links to give color and an active email option. Also changed visited, hover, and link colors. removed body border to clean page up.